Extract Jvm.CrossAppDomain for cleaner code - fixing warnings
